requirements.txt
antlr4-python3-runtime==4.13.0
antlr4-tools
black
cfpq-data
networkx
pre-commit
pydot
pyformlang
pytest
scipy
